chore: refactor the code and add check for input shape
